day13/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.5.2
openai>=1.3.0
numpy==1.24.3
beautifulsoup4==4.12.2
pypdf==3.17.0

